requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
Pillow==10.4.0
pikepdf==8.15.0
pdf2image==1.17.0
gunicorn==21.2.0
